Ensure DeferredQueryExpressions avoid grouping
By formatting a DeferredQueryExpression as a string result, it causes
the objects to be coerced to a str, leading to invalid queries being
generated.

This commit changes the behaviour of convert_condition_group to only
apply precedence grouping to non-deferred filters. This does mean that
any DeferredQueryExpressions must implement precedence themselves.

Fixes issue SigmaHQ/pySigma#69.