Add __hash__ function to `SigmaLevel`
By implementing __eq__ but not __hash__, the `SigmaLevel` enum can no
longer be used as the key of a dictionary (see the second paragraph
of this Python documentation (1)). The code for the ATT&CK framework
integration was however using it in this way (2), and this was leading
to runtime errors in `sigma-cli`. This PR adds in an implementation of
__hash__ to allow it to continue to be used as such, and similarly for
`SigmaStatus`.

1. https://docs.python.org/3.1/reference/datamodel.html#object.__hash__
2. https://github.com/SigmaHQ/sigma-cli/blob/main/sigma/analyze/attack.py#L19