Enable AddCondition to add negated conditions
The AddConditionTransformation currently allows the addition of rule
conditions that are required to be true to a rule, but cannot be used to
add conditions that are negated (e.g., filters).

This adds a "negated" attribute to the transformation, which, if set,
puts a "not" in front of the condition to enable this functionality.