Initial design for fieldref wildcard support
Field references can currently only specify a full match on two fields,
but for backends that support it, we have found that in practice being
able to partially match one field against another is important. This
commit introduces the ability for a field reference to have a leading
or trailing wildcard and updates the contains, startswith, and endswith
modifiers to allow the rule to set them.